requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openpyxl==3.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.1
WTForms==3.2.1
